+++ packages/aikit-respond/src/index.ts
@@ -1,15 +1,7 @@
-export { RespondIoClient, RespondIoClientConfig } from "./respondio";
-
-export {
-  type MessageReceivedPayload,
-  type MessageSentPayload,
-  type ContactAssigneeUpdatedPayload,
-  type ConversationClosedPayload,
-  type WebhookPayload,
-} from "./webhook/types";
-
-export { type RespondIoEvents } from "./webhook/types";
-
-export { RespondIoError, SignatureVerificationError } from "./webhook/errors";
-
+export * from "./webhook/types";
+export * from "./webhook/errors";
 export { RespondIoWebhook } from "./webhook";
+
+export * from "./respondio/types";
+export * from "./respondio/errors";
+export { RespondIoClient } from "./respondio";
